pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evolution-agent"
version = "0.1.0"
description = "Evolutionary code optimization framework with two-tier LLM strategy"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "openai>=1.0",
]

[project.optional-dependencies]
optuna = [
    "optuna>=3.0",
    "cmaes>=0.10",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "optuna>=3.0",
    "cmaes>=0.10",
]

[project.scripts]
evolution-agent = "evolution_agent.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
